- New flag for pop ups "LEAVE_ii", means... do not remove nodes from translate list after unmount( most likely because popup might not be a clone ) - New flag for popups "EXISTS" means... this is a re-render of an existing instance, so just show the window in the DOM - arrHas upgraded... search for more than one item - new method for updating window instance
There was a bug with translations for popup form window, popup window has 2 parts window frame and content, and window is a clone.
So in this case, form is a different instance inside window and when translate was triggered it works only once... because form's nodes to translate were removed from list with window clone.
There was a need of flag that's prevent the unmount function from removing nodes from translation list. And witch that flags it starts to work, but appears memory leak because window is a clone so window's nodes have to be removed from translation list when unmount ( nodes from window frame, not from content in this case  ).
So another flag "EXIST"
With that flag only wrapper node is updated and content remains the same( loaded from old window cache)